# internal/api/openapi.yaml
openapi: "3.0.3"
info:
  title: LLM Interceptor API
  description: Local-first, open-source LLM gateway. Transparent proxy, observability (OTel), governance (budget/rate-limit/tool-policy), LLM routing, and web UI.
  version: "0.3.0"
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /v1/messages:
    post:
      summary: Proxy Anthropic Messages API
      operationId: proxyAnthropic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Successful response
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "402":
          $ref: "#/components/responses/ErrorResponse"
        "429":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /v1/chat/completions:
    post:
      summary: Proxy OpenAI Chat Completions API
      operationId: proxyOpenAI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Successful response
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "402":
          $ref: "#/components/responses/ErrorResponse"
        "429":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
...
components:
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              type:
                type: string
              error:
                type: object
                properties:
                  type:
                    type: string
                  message:
                    type: string